Initrfs build system and test support
