--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "linux"]
-	path = linux
-	url = https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
